Refactor App component to better break down Items component
